Better scene builder format
