+++ JavaSMPwD/src/application/Controller.java
@@ -1,0 +1,9 @@
+package application;
+
+import javafx.event.ActionEvent;
+
+public class Controller {
+	public void saySomething(ActionEvent event) {
+		System.out.println("Hi");
+	}
+}
