+++ JavaSMPwD/src/module-info.java
@@ -4,6 +4,7 @@ module javaSMP {
 	requires javafx.controls;
 	requires javafx.graphics;
 	requires javafx.web;
+	requires javafx.base;
 	
-	opens mainPackage;
+	opens application;
 }
